requirements.txt
# Core Dependencies
pydantic>=2.0.0
pydantic-ai==0.1.10
python-dotenv>=1.0.0
typing-extensions>=4.5.0
more-itertools>=9.1.0

# Agent Framework
asyncio>=3.4.3
aiohttp>=3.8.4

# Supabase Connection
supabase>=2.1.0

# A2A Protocol
a2a-protocol>=0.1.0  # This is a placeholder name, replace with actual package

# Local LLM Integration
ollama>=0.1.5

# Web Crawling
crawl4ai>=0.6.0
requests>=2.28.2

# Data Processing
numpy>=1.24.3
pandas>=2.0.0

# Web Interface
streamlit>=1.22.0

# MCP Integration
mcp-client>=0.3.0
